requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

_libgcc_mutex=0.1=main
absl-py=0.12.0=pyhd8ed1ab_0
aiohttp=3.7.4=py39h27cfd23_1
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=py_1000
attrs=20.3.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pypi_0
blinker=1.4=py_1
brotlipy=0.7.0=py39h38d8fee_1001
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cachetools=4.2.2=pyhd8ed1ab_0
certifi=2020.12.5=py39hf3d152e_1
cffi=1.14.5=py39h261ae71_0
chardet=3.0.4=py39h079e4ff_1008
click=7.1.2=pyh9f0ad1d_0
configargparse=1.4=pypi_0
cryptography=3.4.7=py39hbca0aa6_0
cycler=0.10.0=pypi_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
deprecate=1.0.5=pypi_0
entrypoints=0.3=pypi_0
flask=1.1.2=pypi_0
fsspec=2021.5.0=pypi_0
future=0.18.2=pypi_0
google-auth=1.30.0=pyh44b312d_0
google-auth-oauthlib=0.4.4=pypi_0
grpcio=1.37.1=pypi_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.0.1=py39hf3d152e_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py39hb070fc8_0
ipython=7.22.0=py39hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
itsdangerous=2.0.0=pypi_0
jedi=0.17.2=py39h06a4308_1
jinja2=2.11.3=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py39h06a4308_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=1.1.1=pypi_0
matplotlib=3.4.1=pypi_0
matplotlib2tikz=0.7.6=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py39h27cfd23_1
mkl_fft=1.3.0=py39h42c9631_2
mkl_random=1.2.1=py39ha9443f7_2
multidict=5.1.0=py39h27cfd23_2
nbclient=0.5.3=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
notebook=6.3.0=pypi_0
numpy=1.20.2=pypi_0
numpy-base=1.20.1=py39h7d8b39e_0
oauthlib=3.1.0=pypi_0
opennmt-py=2.1.2=dev_0
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pypi_0
pandas=1.2.4=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=pypi_0
pip=21.0.1=py39h06a4308_0
portalocker=2.0.0=pypi_0
prometheus-client=0.10.1=pypi_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.16.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pydeprecate=0.3.0=pypi_0
pygments=2.8.1=pyhd3eb1b0_0
pyjwt=2.1.0=pyhd8ed1ab_0
pyonmttok=1.26.0=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=py39hf3d152e_3
python=3.9.2=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.9=1_cp39
pytorch-lightning=1.2.0=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4=pypi_0
pyzmq=20.0.0=py39h2531618_1
qtconsole=5.0.3=pypi_0
qtpy=1.9.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
sacrebleu=1.5.1=pypi_0
scipy=1.6.3=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=52.0.0=py39h06a4308_0
six=1.15.0=py39h06a4308_0
sqlite=3.35.4=hdfb4753_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorboardx=2.2=pypi_0
terminado=0.9.4=pypi_0
testpath=0.4.4=pypi_0
tikzplotlib=0.9.8=pypi_0
tk=8.6.10=hbc83047_0
torch=1.8.0+cu102=pypi_0
torch-tb-profiler=0.1.0=pypi_0
torchmetrics=0.3.2=pypi_0
torchtext=0.5.0=pypi_0
torchvision=0.9.0=pypi_0
tornado=6.1=py39h27cfd23_0
tqdm=4.60.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.4=pyhd8ed1ab_0
waitress=1.4.4=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py39hbd71b63_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
